+++ core/docs/changelog.md
@@ -1,6 +1,7 @@
 # TBD
 ### Features
 * Support `CopyFilesFromUserService` for Kubernetes
+* `StartService` and `GetServiceInfo` now return the Service GUID
 
 # 1.50.2
 ### Fixes
@@ -11,7 +12,6 @@
 ### Changes
 * Update files expansion logic to use unified expansion method from container-engine-lib
 * Upgrade to container-engine-lib 0.27.0, which implements the unified file artifact expansion logic
-* `StartService` and `GetServiceInfo` now return the Service GUID
 
 # 1.50.0
 ### Features