+++ core/docs/changelog.md
@@ -1,4 +1,6 @@
 # TBD
+
+# 1.50.3
 ### Features
 * Support `CopyFilesFromUserService` for Kubernetes
 * `StartService` and `GetServiceInfo` now return the Service GUID